requirements.txt
Django==5.2.7
django-allauth==65.11.2
python-decouple==3.8
fontawesomefree==6.6.0
Pillow==11.3.0
psycopg==3.2.10